fix: consolidate imports in tokenization and tokenization_stream examples for cleaner code
